requirements.txt
pydantic-settings==2.10.1
weaviate-client==4.16.9
requests==2.32.5
ipython==9.5.0
pandas==2.3.2
pillow==11.3.0
matplotlib==3.10.6
openai==1.107.3
fastapi==0.116.2
uvicorn[standard]==0.35.0
python-multipart==0.0.20
